Factorize request response handling into a new function request_handle_response().
